Extract fetchBlueprints method to a util
