+++ src/app/steps/Fork/ForkStep.js
@@ -1,49 +1,11 @@
-import { dispatch, select } from '@wordpress/data';
 import { Container, Title } from '@newfold/ui-component-library';
-import { nfdOnboardingStore } from '@/data/store';
 import { Navigate } from '@/components';
 import { OnboardingEvent, trackOnboardingEvent } from '@/utils/analytics/hiive';
 import { ACTION_FORK_OPTION_SELECTED, ACTION_ONBOARDING_STARTED } from '@/utils/analytics/hiive/constants';
-import { disableComingSoon, getBlueprints } from '@/utils/api';
+import { disableComingSoon } from '@/utils/api';
+import { fetchBlueprints } from '@/utils/blueprints';
 import ForkOptions from './ForkOptions';
 import ForkLinks from './ForkLinks';
-
-/**
- * Fetch the blueprints.
- */
-const fetchBlueprints = async () => {
-	// If the blueprints are already fetched, return.
-	const blueprints = select( nfdOnboardingStore ).getBlueprints();
-	if ( blueprints.length > 0 ) {
-		return;
-	}
-
-	// Fetch the blueprints.
-	getBlueprints().then( ( response ) => {
-		if ( ! response || response.error || ! Array.isArray( response.body ) ) {
-			// eslint-disable-next-line no-console
-			console.error( response.error );
-		} else {
-			dispatch( nfdOnboardingStore ).setBlueprints( response.body );
-
-			// Pre-download the preview images to avoid flickering.
-			const images = response.body.map( ( blueprint ) => {
-				return blueprint.screenshot_url;
-			} );
-			Promise.allSettled(
-				images.map( ( imageUrl ) => {
-					return new Promise( ( resolve, reject ) => {
-						// eslint-disable-next-line no-undef
-						const img = new Image();
-						img.onload = () => resolve( imageUrl );
-						img.onerror = () => reject( imageUrl );
-						img.src = imageUrl;
-					} );
-				} )
-			);
-		}
-	} );
-};
 
 const ForkStep = () => {
 	const [ selectedForkOption, setSelectedForkOption ] = useState( null );
