+++ src/app/utils/blueprints/index.js
@@ -1,0 +1,1 @@
+export { default as fetchBlueprints } from './fetchBlueprints';
